+++ backend/apis/api_users/api.py
@@ -1,7 +1,11 @@
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes, authentication_classes
 from rest_framework import status
 
+from rest_framework.permissions import IsAuthenticated
+from rest_framework_simplejwt.authentication import JWTAuthentication
+from apis.permissions_decorators import IsAdmin
+from rest_framework.exceptions import PermissionDenied
 
 from apis.models import User, UserValoration
 from .serializers import UserSerializer
@@ -9,7 +13,20 @@ from apis.api_valoration.serializers import UserValorationSerializer
 
 
 @api_view(['GET', 'POST'])
+@authentication_classes([JWTAuthentication])
 def user_api_view(request):
+
+    def check_permissions(request):
+
+        if request.method == 'GET':
+            for permission_class in [IsAuthenticated, IsAdmin]:
+                permission = permission_class()
+                if not permission.has_permission(request, None):
+                    raise PermissionDenied(getattr(permission, 'message', None))
+                
+    
+    check_permissions(request)
+
 
     if request.method == 'GET':
 
@@ -20,7 +37,12 @@ def user_api_view(request):
     
     #create
     elif request.method == 'POST':
+        
         data = request.data
+        if data.get('isAdministrator'):
+            return Response({'message':'No puede crear un administrador. Contacte con alguno para que este le de permisos de administrador'}, status=status.HTTP_403_FORBIDDEN)
+
+
         serializer = UserSerializer(data = data) #many false es para indicar que realizamos un post a la vez
         if serializer.is_valid():
             serializer.save()
@@ -30,8 +52,33 @@ def user_api_view(request):
 
 
 
+
 @api_view(['GET', 'PUT', 'DELETE'])
+@authentication_classes([JWTAuthentication])
 def user_detail_api_view(request, id):
+
+
+    def check_permissions(request):
+
+        if request.method == 'GET':
+            for permission_class in [IsAuthenticated]:
+                permission = permission_class()
+                if not permission.has_permission(request, None):
+                    raise PermissionDenied(getattr(permission, 'message', None))
+
+        if request.method == 'PUT':
+            for permission_class in [IsAuthenticated]:
+                permission = permission_class()
+                if not permission.has_permission(request, None):
+                    raise PermissionDenied(getattr(permission, 'message', None))
+                
+        elif request.method == 'DELETE':
+            for permission_class in [IsAuthenticated]:
+                permission = permission_class()
+                if not permission.has_permission(request, None):
+                    raise PermissionDenied(getattr(permission, 'message', None))
+
+    check_permissions(request)
         
     # queryset
     user = User.objects.filter(id=id).first()
@@ -40,18 +87,47 @@ def user_detail_api_view(request, id):
     if user:
         
         if request.method == 'GET':
+            if not (request.user.id == id or request.user.isAdministrator):
+                return Response({'message':'Solo puede ver sus datos, no puede ver información ajena'}, status=status.HTTP_403_FORBIDDEN)
+
+
             serializer = UserSerializer(user)
             return Response(serializer.data, status=status.HTTP_200_OK)
         
         elif request.method == 'PUT':
+
+            if not (request.user.id == id or request.user.isAdministrator):
+                return Response({'message':'Solo puede actualizar sus datos, no puede actualizar información ajena'}, status=status.HTTP_403_FORBIDDEN)
+            
+            aux1 = user.isActive
+            aux2 = user.isAdministrator
+
+            if not aux1 and not aux2:
+                return Response({'message':'No puede actualizar su información, su cuenta ha sido desactivada'}, status=status.HTTP_403_FORBIDDEN)
+
+            
+
             data = request.data
+            
+            if data.get('isAdministrator') and aux2 == False and not request.user.isAdministrator:
+                return Response({'message':'No puede hacerse administrador a si mismo'}, status=status.HTTP_403_FORBIDDEN)
+
+            if data.get('isActive') == False and not request.user.isAdministrator:
+                return Response({'message':'No puede desactivar su cuenta sin ser administrador'}, status=status.HTTP_403_FORBIDDEN)
+
+
             serializer = UserSerializer(user, data = data)
             if serializer.is_valid():
                 serializer.save()
                 return Response(serializer.data, status=status.HTTP_200_OK)
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         
+
         elif request.method == 'DELETE':
+
+            if not (request.user.id == id or request.user.isAdministrator):
+                return Response({'message':'Solo puede borrar sus datos, no puede borrar información ajena'}, status=status.HTTP_403_FORBIDDEN)
+            
             user.delete()
             return Response({'message':"Usuario eliminado correctamente!"}, status=status.HTTP_200_OK)
         
@@ -60,6 +136,8 @@ def user_detail_api_view(request, id):
 
 
 @api_view(['GET'])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
 def user_valoration_received_api_view(request, id):
 
     user = User.objects.filter(id=id).first()
@@ -85,6 +163,8 @@ def user_valoration_received_api_view(request, id):
 
 
 @api_view(['GET'])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
 def user_valoration_done_to_users_api_view(request, id):
 
     user = User.objects.filter(id=id).first()
@@ -93,14 +173,11 @@ def user_valoration_done_to_users_api_view(request, id):
 
         if request.method == 'GET':
 
-            #queryset
-            valorations = UserValoration.objects.all()
-            res = []
-            
-            for v in valorations:
-                if v.valuer == user:
-                    res.append(v)
+            if not (request.user.id == id or request.user.isAdministrator):
+                return Response({'message':'Solo puede obtener un listado de todas sus valoraciones, no de otros usuarios'}, status=status.HTTP_403_FORBIDDEN)
 
+            #queryset
+            res = UserValoration.objects.filter(valuer=user)
             serializer = UserValorationSerializer(res, many=True)
             return Response(serializer.data, status = status.HTTP_200_OK)
